Fixed segfault caused by use-after-free
When removing a configuration from the supplicant, the wpa_config
struct will be freed in the process. In Android, a notification
will be sent to user space for this event. However, the message
reads the SSID, which has already been freed. There are two places
pointing to the same SSID struct, and this patch explicitly set
one to NULL after removal of configuration. This fixes the segfault
that may appear when "forgetting" a network in user space.

Issue: AXIA-538
Change-Id: Ia1d62d7c0a9f76f258d0dbba01254c45e07d345d
Signed-off-by: Daniel Leung <daniel.leung@intel.com>