Add a new message type to wpa_supplicant for reporting conditions
Added a CTRL-EVENT-EAP-STATUS message that will be sent to cli or UI when
certain events are generated during an authentication.  This, combined with
the existing CTRL-EVENT-EAP-TLS-CERT-ERROR message type should provide a
fairly clear indication of what portion of the authentication failed, and
why.

Patch behavior modified to meet requirements specified on the wpa_supplicant
mailing list.

Change-Id: I4660fdc77d408828dd5ff6f2d47e06acc0531ff6
Signed-off-by: Chris Hessing <chris.hessing@cloudpath.net>