wpa_supplicant: Add get_capability freq
Add "get_capability freq" command to print a more verbose list of channels and
frequencies in MHz. The tag "NO_IBSS" is added, if IBSS mode is not allowed on
a channel. This allows userspace programs to get the frequencies and
limitations associated with each channel.

This extends the information already provided in "get_capability channel" but a
new interface is chosen because of backwards compatibility considerations.

Change-Id: Ib32c8a1c78cb28933985f70aee7ce4ee2966d074
Signed-hostap: Bruno Randolf <br1@einfach.org>