Return err code when WEP password format is wrong
Originally the wifi driver does not return errer code
when password format of WEP connection is not correct,
but instead directly drop the connection. As a result,
the Android framework does nothing on UI to indicate
that users have input a wrong password, leading to
poor user experience. This change fixes the problem
by returning corresponding error code to inform
Android framework that the given password user input
is in wrong format.

Change-Id: Ib8a4d56221562e5692fcb136f9f534bf79007ba1